refactor(templates): rename model file
