+++ app/portainer/services/api/templateService.js
@@ -1,5 +1,5 @@
 import { commandStringToArray } from '@/docker/helpers/containers';
-import { TemplateViewModel } from '@/react/portainer/templates/app-templates/template';
+import { TemplateViewModel } from '@/react/portainer/templates/app-templates/view-model';
 import { DockerHubViewModel } from 'Portainer/models/dockerhub';
 
 angular.module('portainer.app').factory('TemplateService', TemplateServiceFactory);
