+++ app/react/portainer/templates/app-templates/view-model.ts
